Extract common color coding logic into a helper
This change moves the business logic from Mapbox syntax (yuk) to javascript (yay!). If we change criteria for what constitutes an urgent request we can do it in one place. Likewise, if we decide to change the colors that represent urgency, it happens in one place. And the logic is now protected by tests.

- Use helper in table and map.
- Add tests for helper.