Color-code the map markers according to their age (#154)
* Work-in-progress on #126

This is an attempt at color-coding the map markers according to their age. It doesn't work!

* Color-code the map markers according to their age

Resolves #126

* Fix linting error -- oops!

* Remove work-in-progress code

* Extract common color coding logic into a helper

This change moves the business logic from Mapbox syntax (yuk) to javascript (yay!). If we change criteria for what constitutes an urgent request we can do it in one place. Likewise, if we decide to change the colors that represent urgency, it happens in one place. And the logic is now protected by tests.

- Use helper in table and map.
- Add tests for helper.

* Make map markers more accessible for folks with low vision

Adds a dark gray border around map markers.

Co-authored-by: Emma Smith <emma.smith@reaktor.com>